refactor: put all generators into a folder and separat file
