+++ src/core/built-in/index.ts
@@ -1,5 +1,4 @@
 export * from './build';
 export * from './extend';
-export * from './generators';
 export * from './template';
 export * from './use';
