+++ src/core/generators/time.generator.ts
@@ -1,11 +1,12 @@
 import { dateTime } from './date-time.generator';
+import { DateTimeOpts } from './types';
 
 const defaultTimeFormat = 'HH:mm:ss';
 
 export function time(min: string, max: string) {
-  const timeOpts = {
-    minFormat: defaultTimeFormat,
-    maxFormat: defaultTimeFormat,
+  const timeOpts: DateTimeOpts = {
+    startFormat: defaultTimeFormat,
+    endFormat: defaultTimeFormat,
   };
 
   return dateTime(min, max, timeOpts);
