+++ src/core/index.ts
@@ -1,3 +1,4 @@
 export * from './built-in';
+export * from './generators';
 export * from './types';
 export * from './util';
